requirements.txt
pdfkit
reportlab
requests
flask
python-dotenv
python-pptx
pywin32
pythoncom-binary
langchain
langchain-community
faiss-cpu
ctransformers
sentence-transformers
